[REFACTOR] Update Divide Update User Image Section
유저 이미지 업데이트 영역 분리
